Bahar: feature/grade_rounding (#36)
* rounding the grades

* position change (import math)

---------

+++ src/spz/models.py
@@ -114,7 +114,8 @@ class Attendance(db.Model):
 
     ects_points = db.Column(db.Integer, nullable=False, default=0)
     # internal representation of the grade is in %
-    grade = db.Column(db.Integer, nullable=True)  # TODO store grade encrypted
+    grade = db.Column(db.Float, nullable=True)  # TODO store grade encrypted
+
     # if a student only wants 'bestanden' instead of the grade value, is set to true
     hide_grade = db.Column(db.Boolean, nullable=False, default=False)
